image-recognition/requirements.txt
torch
torchvision
flask
pillow
pymongo
opencv-python
scikit-learn